Continue fetching dashboard stats even when dashboard tab is not active
This change might introduce additional load on the Redis server, but
should be negligible in most cases, and improves user experience by
keeping the dashboard data up to date when switching tabs.

Without it, if the user spends a long time on other tabs, the dashboard
data shows a huge spike the moment they switch back, which is jarring.